Bug#54788 DROP DATABASE can remove hidden blob tables without removing parent tables
1) Drop database code should not attempt to directly drop Blob part tables
   (NDB$BLOB_T_C)
2) NdbApi should not allow Blob part tables to be dropped by applications
   This is true for normal NdbApi usage, but not for the 'Global dict cache' 
   mode used by MySQLD.
3) NdbApi should allow a main table with dropped Blob part tables to be 
   'opened', so that it can be dropped.  Otherwise we are left with an
   indestructable table.

A new testcase is added which reproduces the problem fixed by (1 + 2) above.
There is no automated testcase for fix 3, but it has been tested manually.

+++ storage/ndb/src/ndbapi/NdbDictionary.cpp
@@ -1649,22 +1649,58 @@ int
 NdbDictionary::Dictionary::createTable(const Table & t)
 {
   DBUG_ENTER("NdbDictionary::Dictionary::createTable");
-  DBUG_RETURN(m_impl.createTable(NdbTableImpl::getImpl(t)));
+  if (likely(! is_ndb_blob_table(t.getName())))
+  {
+    DBUG_RETURN(m_impl.createTable(NdbTableImpl::getImpl(t)));
+  }
+  else
+  {
+    /* 4307 : Invalid table name */
+    m_impl.m_error.code = 4307;
+    DBUG_RETURN(-1);
+  }
 }
 
 int
 NdbDictionary::Dictionary::dropTable(Table & t){
-  return m_impl.dropTable(NdbTableImpl::getImpl(t));
+  if (likely(! is_ndb_blob_table(t.getName())))
+  {
+    return m_impl.dropTable(NdbTableImpl::getImpl(t));
+  }
+  else
+  {
+    /* 4249 : Invalid table */
+    m_impl.m_error.code = 4249;
+    return -1;
+  }
 }
 
 int
 NdbDictionary::Dictionary::dropTableGlobal(const Table & t){
-  return m_impl.dropTableGlobal(NdbTableImpl::getImpl(t));
+  if (likely(! is_ndb_blob_table(t.getName())))
+  {
+    return m_impl.dropTableGlobal(NdbTableImpl::getImpl(t));
+  }
+  else
+  {
+    /* 4249 : Invalid table */
+    m_impl.m_error.code = 4249;
+    return -1;
+  }
 }
 
 int
 NdbDictionary::Dictionary::dropTable(const char * name){
-  return m_impl.dropTable(name);
+  if (likely(! is_ndb_blob_table(name)))
+  {
+    return m_impl.dropTable(name);
+  }
+  else
+  {
+    /* 4307 : Invalid table name */
+    m_impl.m_error.code = 4307;
+    return -1;
+  }
 }
 
 bool
